+++ Code/Launcher/CPU.cpp
@@ -10,14 +10,14 @@ bool CPU::IsAMD()
 {
 	const char *vendorID = "AuthenticAMD";  // must be 12 characters long
 
-	int cpuInfo[4];
-	__cpuid(cpuInfo, 0x0);
+	int registers[4];
+	__cpuid(registers, 0x0);
 
 	const int *id = reinterpret_cast<const int*>(vendorID);
 
-	return cpuInfo[1] == id[0]   // 1st part is in EBX register
-	    && cpuInfo[3] == id[1]   // 2nd part is in EDX register
-	    && cpuInfo[2] == id[2];  // 3rd part is in ECX register
+	return registers[1] == id[0]   // 1st part is in EBX register
+	    && registers[3] == id[1]   // 2nd part is in EDX register
+	    && registers[2] == id[2];  // 3rd part is in ECX register
 }
 
 /**
@@ -26,10 +26,10 @@ bool CPU::IsAMD()
  */
 bool CPU::Has3DNow()
 {
-	int cpuInfo[4];
-	__cpuid(cpuInfo, 0x80000001);
+	int registers[4];
+	__cpuid(registers, 0x80000001);
 
-	int bit3DNow = cpuInfo[3] & (1 << 31);  // bit 31 in EDX register
+	int bit3DNow = registers[3] & (1 << 31);  // bit 31 in EDX register
 
 	return bit3DNow != 0;
 }