+++ Code/Launcher/CmdLine.h
@@ -5,6 +5,7 @@
 namespace CmdLine
 {
 	bool HasArg(const char *arg);
+
 	std::string GetArgValue(const char *arg, const char *defaultValue = NULL);
 	int GetArgValueInt(const char *arg, int defaultValue = 0);
 }