+++ Code/Launcher/Launcher.cpp
@@ -6,79 +6,60 @@
 #include <cstring>
 
 #include "CryCommon/IGameStartup.h"
-#include "CryCommon/ILog.h"
 
 #include "Launcher.h"
+#include "CrysisLibs.h"
 #include "CrashLogger.h"
-#include "DLL.h"
-#include "Util.h"
+#include "Error.h"
 
 #include "config.h"
 
-bool Launcher::initCmdLine()
-{
-	const char *cmdLine = Util::GetCmdLine();
-	const size_t cmdLineLength = std::strlen(cmdLine);
+// CryCommon/ISystem.h
+SSystemGlobalEnvironment *gEnv;
 
-	if (cmdLineLength >= sizeof m_params.cmdLine)
+void Launcher::setCmdLine(const char *cmdLine)
+{
+	const size_t length = std::strlen(cmdLine);
+
+	if (length >= sizeof m_params.cmdLine)
 	{
-		Util::ErrorBox("Command line is too long!");
-		return false;
+		throw Error("Command line is too long!");
 	}
 
-	std::memcpy(m_params.cmdLine, cmdLine, cmdLineLength + 1);
-
-	return true;
+	std::memcpy(m_params.cmdLine, cmdLine, length + 1);
 }
 
-void Launcher::logInfo(const char *format, ...)
-{
-	// CryEngine must be initialized
-	ILog *pLog = m_params.pSystem->GetILog();
-
-	va_list args;
-	va_start(args, format);
-	pLog->LogV(ILog::eAlways, format, args);
-	va_end(args);
-}
-
-bool Launcher::run(const DLL & libCryGame)
+void Launcher::run(const CrysisLibs & libs)
 {
 	CrashLogger::Init(m_params.logFileName);
 
-	if (!initCmdLine())
+	IGameStartup::TEntryFunction entry = libs.getCryGame().getSymbol<IGameStartup::TEntryFunction>("CreateGameStartup");
+	if (!entry)
 	{
-		return false;
+		throw Error("The CryGame DLL is not valid!");
 	}
 
-	IGameStartup::TEntryFunction pEntry = libCryGame.getSymbol<IGameStartup::TEntryFunction>("CreateGameStartup");
-	if (!pEntry)
-	{
-		Util::ErrorBox("The CryGame DLL is not valid!");
-		return false;
-	}
-
-	IGameStartup *pGameStartup = pEntry();
+	IGameStartup *pGameStartup = entry();
 	if (!pGameStartup)
 	{
-		Util::ErrorBox("Failed to create the GameStartup Interface!");
-		return false;
+		throw Error("Failed to create the GameStartup Interface!");
 	}
 
 	// initialize CryEngine
 	if (!pGameStartup->Init(m_params))
 	{
-		Util::ErrorBox("Game initialization failed!");
 		pGameStartup->Shutdown();
-		return false;
+		throw Error("Game initialization failed!");
 	}
 
-	logInfo(C1LAUNCHER_VERSION_DESCRIPTION);
+	gEnv = m_params.pSystem->GetGlobalEnvironment();
 
-	// enter game update loop
-	int status = pGameStartup->Run(NULL);
+	CryLogAlways("%s", C1LAUNCHER_VERSION_DESCRIPTION);
+
+	// game update loop
+	m_exitCode = pGameStartup->Run(NULL);
+
+	CryLogAlways("Exit code: %d", m_exitCode);
 
 	pGameStartup->Shutdown();
-
-	return status == 0;
 }