+++ CHANGELOG.md
@@ -5,10 +5,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## [Unreleased]
 ### Added
-- Support for Crysis Wars.
+- Crysis Wars support.
 - Unlocked advantages of pre-ordered version. Both game launcher and server launcher are required for this to work.
 - Crash logger that writes crash info to the log file.
-- Launcher version info is now written to the log after engine initialization.
+- Launcher version can now be found in the log after engine initialization.
 
 ### Fixed
 - No more hangs after crash when running under Wine. Fixed by disabling broken unhandled exception handler in CrySystem.